use IContainerRuntimeProps in aqueduct
