+++ packages/runtime/client-api/package.json
@@ -52,6 +52,7 @@
     "@fluidframework/container-definitions": "^0.28.0",
     "@fluidframework/container-loader": "^0.28.0",
     "@fluidframework/container-runtime": "^0.28.0",
+    "@fluidframework/container-runtime-definitions": "^0.28.0",
     "@fluidframework/core-interfaces": "^0.28.0",
     "@fluidframework/datastore": "^0.28.0",
     "@fluidframework/driver-definitions": "^0.28.0",
